# src/main/resources/api.yaml
swagger: "2.0"
info:
  title: API для банковского приложения
  description: Описание API для банковского приложения
  version: 1.0.0
host: localhost:8080
securityDefinitions:
  JWT:
    type: apiKey
    in: header
    name: Authorization
schemes:
  - https
...
paths:
...
  /accounts/{accountId}/:
    delete:
      tags:
        - Account
      summary: удаление аккаунта
      description: метод для удаления аккаунта
      operationId: "deleteAccount"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - name: "accountId"
          in: "path"
          description: "Account id для удаления"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "not found"
  /holidaysStockExchange/{holidayStockExchangeId}/{holidayId}:
    delete:
      tags:
        - HolidayStockExchange
      summary: удаление выходного на бирже
      description: метод для удаления выходного на бирже
      operationId: "deleteHolidayStockExchange"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - name: "holidayStockExchangeId"
          in: "path"
          description: "HolidayStockExchange id для удаления"
          required: true
          type: "integer"
          format: "int64"
        - name: "holidayId"
          in: "path"
          description: "Holiday id для удаления"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "not found"
...
  /accounts/:
    get:
      tags:
        - Account
      summary: получение всех аккаунтов
      description: метод предназначен для получения списка всех аккаунтов
      operationId: getAccounts
      produces:
        - application/json
      responses:
        '200':
          description: список аккаунтов
          schema:
            uniqueItems: true
            type: array
            items:
              $ref: '#/definitions/Account'
        '500':
          description: внутренняя ошибка приложения
          schema:
            $ref: '#/definitions/Exception'
    post:
      tags:
        - Account
      operationId: postAccount
      summary: создание аккаунта
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Account'
      responses:
        '200':
          description: созданый аккаунт
          schema:
            $ref: '#/definitions/Account'
        '500':
          description: внутренняя ошибка приложения
          schema:
            $ref: '#/definitions/Exception'
...
  /holidaysStockExchange/:
    get:
      tags:
        - HolidayStockExchange
      summary: получение всех выходных на бирже
      description: метод предназначен для получения списка всех выходных биржи
      operationId: getHolidaysStockExchange
      produces:
        - application/json
      responses:
        '200':
          description: список выходных биржи
          schema:
            uniqueItems: true
            type: array
            items:
              $ref: '#/definitions/HolidayStockExchange'
        '500':
          description: внутренняя ошибка приложения
          schema:
            $ref: '#/definitions/Exception'
    post:
      tags:
        - HolidayStockExchange
      operationId: postHolidayStockExchange
      summary: создание выходного для биржи
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/HolidayStockExchange'
      responses:
        '200':
          description: созданый выходной биржи
          schema:
            $ref: '#/definitions/HolidayStockExchange'
        '500':
          description: внутренняя ошибка приложения
          schema:
            $ref: '#/definitions/Exception'
...
definitions:
...
  HolidayStockExchange:
    type: object
    description: выходной на бирже
    required:
      - holidayId
      - stockExchangeId
    properties:
      holidayId:
        type: integer
        description: порядковый номер выходного
      stockExchangeId:
        type: integer
        description: порядковый номер биржи
...
  Account:
    type: object
    description: аккаунт
    required:
      - brokerId
      - cash
      - margin
      - userId
      - rateId
    properties:
      id:
        type: integer
        description: порядковый номер
      brokerId:
        type: integer
        description: порядковый номер брокера
      cash:
        type: number
        description: деньги
      margin:
        type: boolean
        description: разрешена ли маржинальная торговля
      rateId:
        type: integer
        description: номер тарифа
      userId:
        type: integer
        description: номер пользователя
...
  Exception:
    type: object
    description: описание ошибки
    properties:
      field:
        type: array
        description: если ошибка валидации, то здесь поле из запроса, по которому произошла ошибка
        items:
          type: string
      message:
        type: string
        description: текст ошибки
      description:
        type: string
        description: человеко-понятное описание ошибки